refactor: extract shared scan table processing logic
Create src/lib/scan-table-utils.ts with groupScansByImage.
Deduplicate identical processScansForTable logic from
dashboard page and images page.
